requirements.txt
asgiref==3.8.1
asttokens==2.4.1
Brotli==1.1.0
chardet==5.2.0
click==8.1.7
decorator==5.1.1
dj-database-url==2.3.0
Django==5.1.3
django-extensions==3.2.3
djangorestframework==3.15.2
executing==2.1.0
gunicorn==23.0.0
h11==0.14.0
ipdb==0.13.13
ipython==8.29.0
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
reportlab==4.2.5
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.32.1
wcwidth==0.2.13
whitenoise==6.8.2
